Refactor 3d viewer code
