+++ nodestudiojs/src/components/base/components/viewer3d.scss
@@ -4,6 +4,7 @@
 
 
 .drag-handle-viewable-h-yellow {
+    cursor: pointer;
     position: absolute;
     width: 100%;
     height: 2px;
@@ -14,6 +15,7 @@
 }
 
 .drag-handle-viewable-h-green {
+    cursor: pointer;
     position: absolute;
     width: 100%;
     height: 2px;
@@ -24,16 +26,18 @@
 }
 
 .drag-handle-dragable-h {
+    cursor: pointer;
     position: relative;
     width: 100%;
     top: -10px;
-    height: 2px;
+    height: 20px;
     background-color: #0000aa00;
 
 }
 
 
 .drag-handle-viewable-v-purple {
+    cursor: pointer;
     position: absolute;
     width: 2px;
     height: 100%;
@@ -44,6 +48,7 @@
 }
 
 .drag-handle-viewable-v-green {
+    cursor: pointer;
     position: absolute;
     width: 2px;
     height: 100%;
@@ -54,8 +59,9 @@
 }
 
 .drag-handle-dragable-v {
+    cursor: pointer;
     position: relative;
-    width: 2px;
+    width: 20px;
     left: -10px;
     height: 100%;
     background-color: #0000aa00;
